bpf: fix bpf compilation errors
Clang was getting confused compiling stalld.bpf.c due to the source
including a user-space header file: <string.h>. This caused the compile
to fail on some arches where clang did not define an arch-specfic macro
like "__aarch64__", causing the includes to go down a 32-bit path rather
than a 64 bit set of includes.

Remove the header and remove the CLANGARCH workarounds in the Makefile.

Signed-off-by: Clark Williams <williams@redhat.com>