stalld.c: use a more reasonable size for reading /proc/stat
Original definition of STAT_MAX_SIZE was to start at 4096 bytes and
update that if we're doing idle cpu detection (default behavior) by
generating a buffer that is ncpus*page_size in size. The buffer size
is used to allocate a buffer on the stack of the functions that do
idle detection by reading /proc/stat. Unfortunately when you're on an
aarch64 system with 128 cores and the kernel page size is 64k bytes,
you end up with a buffer size of 8Mbytes and that is the stack limit
on most RHEL systems. Instant segfault.

Add a function named calc_stat_max() that reads the current size of
/proc/stat, then rounds it up to the nearest page boundary and adds
a page for padding.

Change variable names from 'sched_stat' to 'proc_stat' to be a bit
less confusing (not reading anythng from /sys/kernel/debug/sched/debug,
only from /proc/stat).

Signed-off-by: Clark Williams <williams@redhat.com>
Resolves: RHEL-40146