change backend default to be queue_tracker (eBPF)
Change default execution mode for stalld to using the queue_tracker
backend, which is an eBPF program that watches task enqueue, dequeue and
switch calls in the kernel. This means we avoid parsing the potentially
huge /sys/kernel/debug/sched/debug file for cpu runqueue information so
that we can detect stalls.

Signed-off-by: Clark Williams <williams@redhat.com>